web: drop NetworkManagerAdapter
* It is not used anymore.
